+++ src/main/java/org/truffleruby/core/range/RangeNodes.java
@@ -275,7 +275,7 @@ public abstract class RangeNodes {
 
     }
 
-    @CoreMethod(names = "step", needsBlock = true, optional = 1, lowerFixnum = 1, returnsEnumeratorIfNoBlock = true)
+    @CoreMethod(names = "step", needsBlock = true, optional = 1, lowerFixnum = 1)
     public abstract static class StepNode extends YieldingCoreMethodNode {
 
         @Child private CallDispatchHeadNode stepInternalCall;