+++ src/main/ruby/truffleruby/core/range.rb
@@ -384,6 +384,11 @@ class Range
   end
 
   private def step_internal(step_size=1, &block) # :yields: object
+
+    if !block_given? && Primitive.object_kind_of?(self.begin, Numeric) && (self.end.nil? || Primitive.object_kind_of?(self.end, Numeric))
+      return Enumerator::ArithmeticSequence.new(self, :step, self.begin, self.end, step_size, self.exclude_end?)
+    end
+
     return to_enum(:step, step_size) do
       validated_step_args = Truffle::RangeOperations.validate_step_size(self.begin, self.end, step_size)
       Truffle::RangeOperations.step_iterations_size(self, *validated_step_args)